backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
# --- Modern AI Stack (Strict Versions) ---
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-google-genai==2.0.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4
# DOWNGRADED to 0.5.3 to fix conflict
chromadb==0.5.3
# --- Tools ---
transformers
torch
sentencepiece
sacremoses
duckduckgo-search
deep-translator
# --- Utils ---
httpx
numpy<2